backend/requirements.txt
fastapi>=0.121.3
uvicorn[standard]==0.21.1
python-dotenv==1.0.0
supabase>=2.0.0
pypdf==3.13.0
openai==0.28
httpx>=0.26,<0.29
python-multipart==0.0.6
pydantic>=2.11.7,<3.0.0
